backend/requirements.txt
alembic==1.15.1
annotated-types==0.7.0
anyio==4.9.0
bcrypt==4.3.0
click==8.1.8
ecdsa==0.19.1
email-validator==2.1.1
exceptiongroup==1.2.2
fastapi==0.115.1
greenlet==3.1.1
h11==0.14.0
httptools==0.6.1
idna==3.10
Mako==1.3.9
MarkupSafe==3.0.2
orjson==3.10.7
passlib==1.7.4
psycopg2-binary==2.9.10
pyasn1==0.6.1
pydantic==2.10.6
pydantic-settings==2.2.1
pydantic_core==2.27.2
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
python-jose==3.3.0
python-multipart==0.0.9
pytz==2024.1
PyYAML==6.0.1
rsa==4.9
six==1.17.0
sniffio==1.3.1
SQLAlchemy==2.0.27  
typing_extensions==4.12.2
ujson==5.10.0
uvicorn==0.34.0
uvloop==0.19.0
watchfiles==0.24.0
websockets==12.0
starlette==0.38.0
requests>=2.32.0